Refactor css files
