+++ src/dotnet/APIView/APIViewWeb/Client/css/pages/delete.scss
@@ -1,0 +1,2 @@
+﻿body {
+}
