+++ src/dotnet/APIView/APIViewWeb/Client/css/pages/index.scss
@@ -1,0 +1,9 @@
+@import "../shared/layout.scss";
+@import "../shared/theme-colors.scss";
+@import "../shared/mixins.scss";
+
+#create-review-button {
+    @include bottom-right-floating;
+    @include btn-circle;
+    @include btn-circle-xl;
+}
