+++ src/dotnet/APIView/APIViewWeb/Client/css/pages/legacy-review.scss
@@ -1,0 +1,2 @@
+﻿body {
+}
