+++ src/dotnet/APIView/APIViewWeb/Client/css/pages/profile.scss
@@ -1,0 +1,2 @@
+﻿body {
+}
