+++ src/dotnet/APIView/APIViewWeb/Client/css/pages/revisions.scss
@@ -1,0 +1,7 @@
+@import "../shared/mixins.scss";
+
+#add-revision-button {
+    @include bottom-right-floating;
+    @include btn-circle;
+    @include btn-circle-xl;
+}
