+++ src/dotnet/APIView/APIViewWeb/Client/css/pages/samples.scss
@@ -1,0 +1,7 @@
+@import "../shared/mixins.scss";
+
+#add-sample-button {
+    @include bottom-right-floating;
+    @include btn-circle;
+    @include btn-circle-xl;
+}
