+++ src/dotnet/APIView/APIViewWeb/Client/css/shared/_layout.scss
@@ -1,0 +1,3 @@
+#apiview-logo {
+    width: 30px;
+}
