+++ src/dotnet/APIView/APIViewWeb/Client/css/shared/mixins.scss
@@ -1,0 +1,26 @@
+@import "theme-colors.scss";
+
+@mixin bottom-right-floating {
+    position: fixed;
+    z-index: 10;
+    bottom: 10%;
+    right: 5%;
+}
+
+@mixin btn-circle {
+    width: 45px;
+    height: 45px;
+    line-height: 45px;
+    text-align: center;
+    padding: 0;
+    border-radius: 50%;
+}
+
+@mixin btn-circle-xl {
+    width: 100px;
+    height: 100px;
+    line-height: 1.2rem;
+    font-weight: bold;
+    font-size: 1.0rem;
+    box-shadow: var(--box-shadow-lg);
+}
