+++ src/dotnet/APIView/APIViewWeb/Client/css/shared/theme-colors.scss
@@ -1,0 +1,23 @@
+@import "../../node_modules/bootstrap/scss/functions";
+@import "../../node_modules/bootstrap/scss/variables";
+
+.light-theme {
+    --box-shadow: 0 .5rem 1rem rgba($black, .15);
+    --box-shadow-sm: 0 .125rem .25rem rgba($black, .075);
+    --box-shadow-lg: 0 1rem 3rem rgba($black, .175);
+    --box-shadow-inset: inset 0 1px 2px rgba($black, .075);
+}
+
+.dark-theme {
+    --box-shadow: 0 .5rem 1rem rgba($white, .15);
+    --box-shadow-sm: 0 .125rem .25rem rgba($white, .075);
+    --box-shadow-lg: 0 1rem 3rem rgba($white, .175);
+    --box-shadow-inset: inset 0 1px 2px rgba($white, .075);
+}
+
+.dark-solarized-theme {
+    --box-shadow: 0 .5rem 1rem rgba($gray-100, .15);
+    --box-shadow-sm: 0 .125rem .25rem rgba($gray-100, .075);
+    --box-shadow-lg: 0 1rem 3rem rgba($gray-100, .175);
+    --box-shadow-inset: inset 0 1px 2px rgba($gray-100, .075);
+}
