+++ src/dotnet/APIView/APIViewWeb/Client/package.json
@@ -31,7 +31,8 @@
     "ts-loader": "^6.2.0",
     "typescript": "^3.5.3",
     "webpack": "^5.70.0",
-    "webpack-cli": "^4.9.2"
+    "webpack-cli": "^4.9.2",
+    "bootstrap": "5.0.2"
   },
   "eslintConfig": {
     "root": true,
