+++ src/dotnet/APIView/APIViewWeb/Client/webpack.config.js
@@ -9,22 +9,17 @@ module.exports = {
     revisions: './src/revisions.ts',
     fileInput: './src/file-input.ts',
     review: './src/review.ts',
-    reviews: './src/reviews.ts',
+    index: './src/index.ts',
     userProfile: './src/user-profile.ts',
-    site: './css/site.scss',
-    c: './css/c.scss',
-    cplusplus: './css/cplusplus.scss',
-    csharp: './css/csharp.scss',
-    go: './css/go.scss',
-    java: './css/java.scss',
-    javascript: './css/javascript.scss',
-    json: './css/json.scss',
-    kotlin: './css/kotlin.scss',
-    python: './css/python.scss',
-    swagger: './css/swagger.scss',
-    swift: './css/swift.scss',
-    xml: './css/xml.scss',
-    usagesample: './css/usageSample.scss'
+    conversation: './css/pages/conversation.scss',
+    delete: './css/pages/delete.scss',
+    index: './css/pages/index.scss',
+    legacyReview: './css/pages/legacy-review.scss',
+    profile: './css/pages/profile.scss',
+    requestedReviews: './css/pages/requested-reviews.scss',
+    review: './css/pages/review.scss',
+    revisions: './css/pages/revisions.scss',
+    samples: './css/pages/samples.scss',
   },
   devtool: 'source-map',
   module: {
