+++ src/dotnet/APIView/APIViewWeb/Pages/Assemblies/ReviewLayout.cshtml
@@ -1,17 +1,10 @@
-﻿@section Scripts
-{
-    <script type="text/javascript" src="@Url.Content("~/review.js")" asp-append-version="true"></script>
-    <script type="text/javascript" src="@Url.Content("~/comments.js")" asp-append-version="true"></script>
-    <script type="text/javascript" src="@Url.Content("~/fileInput.js")" asp-append-version="true"></script>
-    <script type="text/javascript" src="@Url.Content("~/revisions.js")" asp-append-version="true"></script>
-}
-@section Styles
-{
-    <link rel="stylesheet" href="~/usagesample.css" asp-append-version="true">
-}
 @{
     Layout = "Shared/_Layout";
     var id = ViewContext.RouteData.Values["id"];
+}
+@{
+    IgnoreSection("Scripts");
+    IgnoreSection("Styles");
 }
 
 <nav class="navbar-light shadow-sm bg-light">
