+++ src/dotnet/APIView/APIViewWeb/Pages/Assemblies/Samples.cshtml
@@ -8,6 +8,15 @@
     var userPreference = PageModelHelpers.GetUserPreference(Model._preferenceCache, User);
     TempData["UserPreference"] = userPreference;
 }
+@section Scripts
+{
+    <script type="text/javascript" src="@Url.Content("~/comments.js")" asp-append-version="true"></script>
+    <script type="text/javascript" src="@Url.Content("~/fileInput.js")" asp-append-version="true"></script>
+}
+@section Styles
+{
+    <link rel="stylesheet" href="~/samples.css" asp-append-version="true">
+}
 
 @if (Model.Samples == null || Model.Sample.FileId == "Bad Deployment" || Model.Sample.FileId == "File Content Missing") // If the samples container does not exist
 {
@@ -41,7 +50,7 @@ else
             </div>
             <div class="row">
                 <div class="col-lg-12 px-0" data-review-id=@Model.Review.ReviewId>
-                    <div class="bottom-right-floating">
+                    <div id="add-sample-button">
                         <button type="button" class="btn btn-primary btn-circle btn-circle-xl" data-toggle="modal" data-target="#uploadModel"><small>ADD</small><br><i class="fas fa-sm fa-plus"></i><br><small>SAMPLE</small></button>
                     </div>
                     @if (Model.SampleContent.Length == 0)
