+++ src/dotnet/APIView/APIViewWeb/Pages/Shared/_Layout.cshtml
@@ -11,17 +11,14 @@
         <link rel="stylesheet" href="~/lib/bootstrap/dist/css/bootstrap.css" />
     </environment>
     <environment exclude="Development">
-        <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
+        <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css"
               asp-fallback-href="~/lib/bootstrap/dist/css/bootstrap.min.css"
               asp-fallback-test-class="sr-only" asp-fallback-test-property="position" asp-fallback-test-value="absolute"
               crossorigin="anonymous"
-              integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" />
+              integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" />
     </environment>
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css" />
     <link rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/4.1.1/animate.min.css"/>
-    <!-- BootStrap Select -->
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css"/>
-    <link id="theme-css" rel="stylesheet" href="~/site.css" asp-append-version="true" />
     @{
         var modelType = ViewData.ModelMetadata.ModelType.Name;
         if (modelType == "ReviewPageModel")
@@ -41,25 +38,19 @@
 </head>
 @{
     var userPreference = (TempData["UserPreference"] as UserPreferenceModel) ?? new UserPreferenceModel();
-    List<string> themes = new List<string>()
-    {
-        "light-theme",
-        "dark-theme",
-        "dark-solarized-theme"
-    };
     var theme = userPreference.Theme;
 }
 <body class="@theme">
     <header>
         <nav class="navbar navbar-expand-sm navbar-toggleable-sm navbar-dark bg-dark">
-            <div class="container-fluid px-0">
-                <a href="/" class="navbar-brand"><img class="apiview-logo" alt="apiview-logo" src="~/icons/apiview.png" />apiview.dev</a>
+            <div class="container-fluid">
+                <a href="/" class="navbar-brand"><img id="apiview-logo" alt="apiview-logo" src="~/icons/apiview.png" />apiview.dev</a>
                 <button class="navbar-toggler" type="button" data-toggle="collapse" data-target=".navbar-collapse" aria-controls="navbarSupportedContent"
                         aria-expanded="false" aria-label="Toggle navigation">
                     <span class="navbar-toggler-icon"></span>
                 </button>
                 <div class="navbar-collapse collapse">
-                    <ul class="navbar-nav mr-auto">
+                    <ul class="navbar-nav me-auto">
                         @if (User.Identity.IsAuthenticated) {
                             <li class="nav-item">
                                 <a class="nav-link" asp-area="" asp-page="/Assemblies/Index">Reviews</a>
@@ -71,7 +62,7 @@
                             </li>
                         }
                     </ul>
-                    <ul class="navbar-nav ml-auto">
+                    <ul class="navbar-nav ms-auto">
                         @if (User.Identity.IsAuthenticated) {
                             <li class="nav-item">
                                 <a asp-page="Profile" asp-route-UserName="@User.GetGitHubLogin()" class="nav-link">Profile</a>
